# ClaimsDesk.API/DTOs/v1/BranchDto.cs
namespace ClaimsDesk.API.DTOs.v1
{
    public class BranchDto
    {
        public int Id { get; set; }
        public string Name { get; set; } = string.Empty;
        public string Code { get; set; } = string.Empty;
        public bool IsHeadOffice { get; set; }
        public string Address { get; set; } = string.Empty;
        public string Status { get; set; } = string.Empty;
        public int OrganizationId { get; set; }
    }

    public class CreateBranchDto
    {
        public string Name { get; set; } = string.Empty;
        public string Code { get; set; } = string.Empty;
        public bool IsHeadOffice { get; set; }
        public string Address { get; set; } = string.Empty;
        public string Status { get; set; } = "Active";
        public int OrganizationId { get; set; }
    }

    public class UpdateBranchDto
    {
        public string Name { get; set; } = string.Empty;
        public string Code { get; set; } = string.Empty;
        public bool IsHeadOffice { get; set; }
        public string Address { get; set; } = string.Empty;
        public string Status { get; set; } = string.Empty;
        public int OrganizationId { get; set; }
    }
}